requirements.txt
streamlit==1.28.1
pyyaml==6.0.1
python-dotenv==1.0.0
azure-identity==1.15.0
azure-mgmt-compute==30.6.0
azure-monitor-query==1.4.0
numpy==1.24.3
pandas==2.0.3
